+++ src/lib/djerba/core/main.py
@@ -7,31 +7,31 @@ Main class to:
 from configparser import ConfigParser
 import json
 import logging
+import pdfkit
 import os
 import re
+from PyPDF2 import PdfMerger
 import djerba.util.ini_fields as ini
 import djerba.version as version
 from djerba.core.base import base as core_base
-from djerba.core.configure import configurer as core_configurer
 from djerba.core.database.archiver import archiver
-from djerba.core.extract import extractor as core_extractor
+from djerba.core.extract import extraction_setup
 from djerba.core.json_validator import plugin_json_validator
-from djerba.core.render import renderer as core_renderer
+from djerba.core.render import html_renderer, pdf_renderer
 from djerba.core.loaders import \
     plugin_loader, merger_loader, helper_loader, core_config_loader
 from djerba.core.workspace import workspace
 from djerba.util.logger import logger
 from djerba.util.validator import path_validator
-import djerba.core.constants as core_constants
+import djerba.core.constants as cc
 import djerba.util.constants as constants
 
 class main(core_base):
 
-    # TODO move to constants file(s)
     PLUGINS = 'plugins'
     MERGERS = 'mergers'
     MERGE_INPUTS = 'merge_inputs'
-    
+
     def __init__(self, work_dir=None, log_level=logging.INFO, log_path=None):
         self.log_level = log_level
         self.log_path = log_path
@@ -48,9 +48,62 @@ class main(core_base):
         self.plugin_loader = plugin_loader(self.log_level, self.log_path)
         self.merger_loader = merger_loader(self.log_level, self.log_path)
         self.helper_loader = helper_loader(self.log_level, self.log_path)
-
     def _get_render_priority(self, plugin_data):
-        return plugin_data[core_constants.PRIORITIES][core_constants.RENDER]
+        return plugin_data[cc.PRIORITIES][cc.RENDER]
+
+    def _load_component(self, name):
+        if name == ini.CORE:
+            component = self.core_config_loader.load(self.workspace)
+        elif self._is_helper_name(name):
+            component = self.helper_loader.load(name, self.workspace)
+        elif self._is_merger_name(name):
+            component = self.merger_loader.load(name)
+        else:
+            component = self.plugin_loader.load(name, self.workspace)
+        return component
+
+    def _resolve_configure_dependencies(self, config, components, ordered_names):
+        self._resolve_ini_deps(cc.DEPENDS_CONFIGURE, config, components, ordered_names)
+
+    def _resolve_extract_dependencies(self, config, components, ordered_names):
+        self._resolve_ini_deps(cc.DEPENDS_EXTRACT, config, components, ordered_names)
+
+    def _resolve_ini_deps(self, depends_key, config, components, ordered_names):
+        for name in ordered_names:
+            if config.has_option(name, depends_key):
+                depends_str = config.get(name, depends_key)
+            else:
+                depends_str = components[name].get_reserved_default(depends_key)
+            depends = self._parse_comma_separated_list(depends_str)
+            if len(depends)>0:
+                failed = 0
+                for dependency in depends:
+                    if self._is_merger_name(dependency):
+                        msg = "Cannot specify dependency on merger '{0}'".format(dependency)
+                        self.logger.error(msg)
+                        raise DjerbaDependencyError(msg)
+                    dependency_ok = False
+                    for other_name in ordered_names:
+                        if other_name == name:
+                            break
+                        elif other_name == dependency:
+                            dependency_ok = True
+                            break
+                    if not dependency_ok:
+                        failed += 1
+                if failed==0:
+                    template = 'Resolved {0} dependencies for component {1}'
+                    self.logger.debug(template.format(len(depends), name))
+                else:
+                    template = 'Failed to resolve {0} of {1} dependencies in {2} for '+\
+                        'component {3}. One or more dependencies is missing or in the '+\
+                        'wrong order. Dependencies: {4} Priority order: {5}'
+                    args = [failed, len(depends), depends_key, name, depends, ordered_names]
+                    msg = template.format(*args)
+                    self.logger.error(msg)
+                    raise DjerbaDependencyError(msg)
+            else:
+                self.logger.debug("No dependencies found for component {0}".format(name))
 
     def _run_merger(self, merger_name, data):
         """Assemble inputs for the named merger and run merge/dedup to get HTML"""
@@ -74,31 +127,31 @@ class main(core_base):
         if config_path_out:  # do this *before* taking the time to generate output
             self.path_validator.validate_output_file(config_path_out)
         config_in = self.read_ini_path(config_path_in)
-        # TODO first read defaults, then overwrite
         components = {}
+        priorities = {}
+        # 1. Load components, set priorities, resolve dependencies (if any)
         self.logger.debug('Loading components and finding config priority levels')
-        for section_name in config_in.sections():
-            if section_name == ini.CORE:
-                component = self.core_config_loader.load()
-            elif self._is_helper_name(section_name):
-                component = self.helper_loader.load(section_name, self.workspace)
-            elif self._is_merger_name(section_name):
-                component = self.merger_loader.load(section_name)
+        for section in config_in.sections():
+            components[section] = self._load_component(section)
+            # if input has a configure priority, use that
+            # otherwise, use default priority for the component
+            if config_in.has_option(section, cc.CONFIGURE_PRIORITY):
+                priority = config_in.getint(section, cc.CONFIGURE_PRIORITY)
             else:
-                component = self.plugin_loader.load(section_name, self.workspace)
-            if config_in.has_option(section_name, core_constants.CONFIGURE_PRIORITY):
-                priority = \
-                    config_in.getint(section_name, core_constants.CONFIGURE_PRIORITY)
-            else:
-                priority = component.get_default_config_priority()
-            components[section_name] = (component, priority)
+                priority = components[section].get_reserved_default(cc.CONFIGURE_PRIORITY)
+            priorities[section] = priority
         self.logger.debug('Configuring components in priority order')
+        ordered_names = sorted(priorities.keys(), key=lambda x: priorities[x])
+        self._resolve_configure_dependencies(config_in, components, ordered_names)
+        # 2. Validate and run configuration for each component; store in config_out
         config_out = ConfigParser()
         order = 0
-        for name in sorted(components, key=lambda x: components[x][1]):
+        for name in ordered_names:
             order += 1
-            component = components[name][0]
-            self.logger.debug('Configuring component {0} in order {1}'.format(name, order))
+            component = components[name]
+            priority = priorities[name]
+            msg = 'Configuring {0}, priority {1}, order {2}'.format(name, priority, order)
+            self.logger.debug(msg)
             component.validate_minimal_config(config_in)
             config_tmp = component.configure(config_in)
             component.validate_full_config(config_tmp)
@@ -116,25 +169,26 @@ class main(core_base):
         if json_path:  # do this *before* taking the time to generate output
             self.path_validator.validate_output_file(json_path)
         components = {}
-        for section_name in config.sections():
-            if section_name == ini.CORE or self._is_merger_name(section_name):
-                continue
-            elif self._is_helper_name(section_name):
-                component = self.helper_loader.load(section_name, self.workspace)
-            else:
-                component = self.plugin_loader.load(section_name, self.workspace)
-            priority = config.getint(section_name, core_constants.EXTRACT_PRIORITY)
-            components[section_name] = (component, priority)
+        priorities = {}
+        # 1. Load components, set priorities, resolve dependencies (if any)
+        for section in config.sections():
+            if not (section == ini.CORE or self._is_merger_name(section)):
+                components[section] = self._load_component(section)
+                priorities[section] = config.getint(section, cc.EXTRACT_PRIORITY)
+        self.logger.debug('Configuring components in priority order')
+        ordered_names = sorted(priorities.keys(), key=lambda x: priorities[x])
+        self._resolve_extract_dependencies(config, components, ordered_names)
+        # 2. Validate and run configuration for each component; store in data structure
         self.logger.debug('Generating core data structure')
-        data = core_extractor(self.log_level, self.log_path).run(config)
+        data = extraction_setup(self.log_level, self.log_path).run(config)
         self.logger.debug('Running extraction for plugins and mergers in priority order')
         order = 0
-        for name in sorted(components, key=lambda x: components[x][1]):
+        for name in ordered_names:
             order += 1
-            component = components[name][0]
+            component = components[name]
             self.logger.debug('Extracting component {0} in order {1}'.format(name, order))
             component.validate_full_config(config)
-            component_data = components[name][0].extract(config)
+            component_data = components[name].extract(config)
             if not self._is_helper_name(name):
                 # only plugins, not helpers, write data in the JSON document
                 self.json_validator.validate_data(component_data)
@@ -156,36 +210,52 @@ class main(core_base):
         self.logger.info('Finished Djerba extract step')
         return data
 
-    def render(self, data, html_path=None):
+    def render(self, data, out_dir=None, pdf=False):
         self.logger.info('Starting Djerba render step')
-        if html_path:  # do this *before* taking the time to generate output
-            self.path_validator.validate_output_file(html_path)
-        body = {} # HTML strings to make up the report body
+        if out_dir:  # do this *before* taking the time to generate output
+            self.path_validator.validate_output_dir(out_dir)
+        html = {} # HTML strings to make up the report file(s)
         priorities = {}
         attributes = {}
+        # 1. Run plugins and mergers to render HTML
         self.logger.debug('Rendering plugin HTML')
         for plugin_name in data[self.PLUGINS]:
-            # render plugin HTML, and find which mergers it uses
             plugin_data = data[self.PLUGINS][plugin_name]
             plugin = self.plugin_loader.load(plugin_name, self.workspace)
-            body[plugin_name] = plugin.render(plugin_data)
+            html[plugin_name] = plugin.render(plugin_data)
             self.logger.debug("Ran plugin '{0}' for rendering".format(plugin_name))
             priorities[plugin_name] = self._get_render_priority(plugin_data)
-            attributes[plugin_name] = plugin_data[core_constants.ATTRIBUTES]
+            attributes[plugin_name] = plugin_data[cc.ATTRIBUTES]
         for (merger_name, merger_config) in data[self.MERGERS].items():
-            body[merger_name] = self._run_merger(merger_name, data)
+            html[merger_name] = self._run_merger(merger_name, data)
             self.logger.debug("Ran merger '{0}' for rendering".format(merger_name))
-            priorities[merger_name] = merger_config[core_constants.RENDER_PRIORITY]
-            attributes[merger_name] = merger_config[core_constants.ATTRIBUTES]
-        self.logger.debug("Assembling HTML document")
-        renderer = core_renderer(self.log_level, self.log_path)
-        html = renderer.run(body, priorities, attributes, data) # TODO remove data argument
-        if html_path:
-            with open(html_path, 'w') as out_file:
-                out_file.write(html)
-            self.logger.info("Wrote HTML output to {0}".format(html_path))
+            priorities[merger_name] = merger_config[cc.RENDER_PRIORITY]
+            attributes[merger_name] = merger_config[cc.ATTRIBUTES]
+        # 2. Assemble plugin/merger HTML outputs according to their priorities/attributes
+        self.logger.debug("Assembling HTML document(s)")
+        h_rend = html_renderer(data[cc.CORE], self.log_level, self.log_path)
+        output_data = h_rend.run(html, priorities, attributes)
+        # 3. Write output files, if any
+        if out_dir:
+            p_rend = pdf_renderer(self.log_level, self.log_path)
+            for prefix in output_data[cc.DOCUMENTS].keys():
+                html_path = os.path.join(out_dir, prefix+'.html')
+                with open(html_path, 'w') as out_file:
+                    out_file.write(output_data[cc.DOCUMENTS][prefix])
+                self.logger.info("Wrote HTML output to {0}".format(html_path))
+                if pdf:
+                    pdf_path = os.path.join(out_dir, prefix+'.pdf')
+                    footer = output_data[cc.PAGE_FOOTER]
+                    p_rend.render_file(html_path, pdf_path, footer)
+                    self.logger.info("Wrote PDF output to {0}".format(pdf_path))
+            merge_list = output_data[cc.MERGE_LIST]
+            if pdf and len(merge_list)>1:
+                merge_in = [os.path.join(out_dir, x+'.pdf') for x in merge_list]
+                merge_out = os.path.join(out_dir, output_data[cc.MERGED_FILENAME])
+                p_rend.merge_pdfs(merge_in, merge_out)
+                self.logger.info("Wrote merged PDF output to {0}".format(merge_out))
         self.logger.info('Finished Djerba render step')
-        return html
+        return output_data
 
     def read_ini_path(self, ini_path):
         self.path_validator.validate_input_file(ini_path)
@@ -196,7 +266,7 @@ class main(core_base):
     def run(self, args):
         # run from command-line args
         # path validation was done in command-line script
-        ap = arg_processor(args, validate=False)
+        ap = arg_processor(args, self.logger, validate=False)
         mode = ap.get_mode()
         work_dir = ap.get_work_dir()
         if mode == constants.CONFIGURE:
@@ -209,25 +279,23 @@ class main(core_base):
             archive = ap.is_archive_enabled()
             config = self.read_ini_path(ini_path)
             self.extract(config, json_path, archive)
-        elif mode == constants.HTML:
+        elif mode == constants.RENDER:
             json_path = ap.get_json_path()
-            html_path = ap.get_html_path()
             with open(json_path) as json_file:
                 data = json.loads(json_file.read())
-            self.render(data, html_path)
+            self.render(data, ap.get_out_dir(), ap.is_pdf_enabled())
         elif mode == constants.REPORT:
             # get operational parameters
             ini_path = ap.get_ini_path()
-            ini_path_out = ap.get_ini_out_path() # may be None
-            json_path = ap.get_json_path() # may be None
-            html_path = ap.get_html_path()
-            #pdf_path = ap.get_pdf()
+            out_dir = ap.get_out_dir()
+            ini_path_out = os.path.join(out_dir, 'full_config.ini')
+            json_path = os.path.join(out_dir, 'djerba_report.json')
             # caching and cleanup are plugin-specific, should be configured in INI
             # can also have a script to auto-populate INI files in 'setup' mode
             archive = ap.is_archive_enabled()
             config = self.configure(ini_path, ini_path_out)
             data = self.extract(config, json_path, archive)
-            self.render(data, html_path)
+            self.render(data, ap.get_out_dir())
             # TODO if pdf_path!=None, convert HTML->PDF
         else:
             # TODO add clauses for setup, pdf, etc.
@@ -241,14 +309,19 @@ class arg_processor(logger):
 
     DEFAULT_JSON_FILENAME = 'djerba_report.json'
 
-    def __init__(self, args, validate=True):
+    def __init__(self, args, logger=None, validate=True):
         self.args = args
-        self.log_level = self.get_args_log_level(self.args)
-        self.log_path = self.args.log_path
-        if self.log_path:
-            # we are verifying the log path, so don't write output there yet
-            path_validator(self.log_level).validate_output_file(self.log_path)
-        self.logger = self.get_logger(self.log_level, __name__, self.log_path)
+        if logger:
+            # do not call 'get_logger' if one has already been configured
+            # this way, we can preserve the level/path of an existing logger
+            self.logger = logger
+        else:
+            self.log_level = self.get_args_log_level(self.args)
+            self.log_path = self.args.log_path
+            if self.log_path:
+                # we are verifying the log path, so don't write output there yet
+                path_validator(self.log_level).validate_output_file(self.log_path)
+            self.logger = self.get_logger(self.log_level, __name__, self.log_path)
         if validate:
             self.validate_args(self.args)  # checks subparser and args are valid
         self.mode = self.args.subparser_name
@@ -287,23 +360,11 @@ class arg_processor(logger):
     def get_log_path(self):
         return self.log_path
 
-    def get_html_path(self):
-        if hasattr(self.args, 'html') and self._get_arg('html') != None:
-           value = self._get_arg('html')
-        else:
-            work_dir = self.get_work_dir()
-            if work_dir == None:
-                msg = "Cannot find default HTML path, work_dir undefined"
-                self.logger.error(msg)
-                raise RuntimeError(msg)
-            value = os.path.join(work_dir, 'djerba_report.html')
-        return value
-
     def get_mode(self):
         return self.mode
 
-    def get_pdf_path(self):
-        return self._get_arg('pdf')
+    def get_out_dir(self):
+        return self._get_arg('out_dir')
 
     def get_work_dir(self):
         # default to None if work_dir is not in args
@@ -320,6 +381,9 @@ class arg_processor(logger):
     def is_cleanup_enabled(self):
         # use to auto-populate INI in 'setup' mode
         return not self._get_arg('no_cleanup')
+
+    def is_pdf_enabled(self):
+        return self._get_arg('pdf')
 
     def validate_args(self, args):
         """
@@ -339,25 +403,13 @@ class arg_processor(logger):
             v.validate_output_dir(args.work_dir)
             if args.json:
                 v.validate_output_file(args.json)
-        elif args.subparser_name == constants.HTML:
+        elif args.subparser_name == constants.RENDER:
             v.validate_input_file(args.json)
-            v.validate_output_file(args.html)
-            if args.pdf:
-                v.validate_output_file(args.pdf)
-        elif args.subparser_name == constants.PDF:
-            v.validate_input_file(args.json)
-            v.validate_output_dir(args.pdf)
+            v.validate_output_dir(args.out_dir)
         elif args.subparser_name == constants.REPORT:
             v.validate_input_file(args.ini)
             v.validate_output_dir(args.work_dir)
-            if args.ini_out:
-                v.validate_output_file(args.ini_out)
-            if args.html:
-                v.validate_output_file(args.html)
-            if args.json:
-                v.validate_output_file(args.json)
-            if args.pdf:
-                v.validate_output_file(args.pdf)
+            v.validate_output_dir(args.out_dir)
         else:
             # shouldn't happen, but handle this case for completeness
             raise ValueError("Unknown subparser: " + args.subparser_name)
@@ -366,5 +418,5 @@ class arg_processor(logger):
 class ArgumentNameError(Exception):
     pass
 
-class ComponentNameError(Exception):
+class DjerbaDependencyError(Exception):
     pass