+++ src/lib/djerba/core/workspace.py
@@ -26,10 +26,13 @@ class workspace(logger):
 
     def abs_path(self, rel_path):
         """Return the absolute path of a file in the workspace"""
-        return os.path.join(self.dir_path, rel_path)
+        return os.path.abspath(os.path.join(self.dir_path, rel_path))
 
     def get_work_dir(self):
         return self.dir_path
+
+    def has_file(self, rel_path):
+        return os.path.exists(os.path.join(self.dir_path, rel_path))
 
     def open_gzip_file(self, rel_path, write=False):
         if write:
@@ -67,6 +70,9 @@ class workspace(logger):
             content = in_file.read()
         return content
 
+    def remove_file(self, rel_path):
+        os.remove(os.path.join(self.dir_path, rel_path))
+    
     # no need to validate paths for write_* methods; output dir already validated as writable
 
     def write_json(self, rel_path, data):