+++ CHANGELOG.md
@@ -1,5 +1,42 @@
 # CHANGELOG
 
+## v1.0.0-dev0.0.7: 2023-08-02
+
+### GCGI-963: Patient info plugin
+- Simple plugin to generate the Clinical Research Report header and Case Overview section
+
+### GCGI-982: Provenance helper update
+- Update to complement changes to core functionality
+- Writes subset of provenance and `sample_info.json` at both configure and extract
+
+## v1.0.0-dev0.0.6: 2023-07-20
+
+### GCGI-967: Overhaul core functionality
+- Define core INI parameters and implement in `core_configurer`
+- Get rid of placeholder data at the core extract step
+- Add PDF rendering to the core
+- Introduce `document_config.json` with settings to render HTML
+- Render multiple HTML/PDF documents, identified by attributes (clinical, research, etc)
+- Add a `mako_renderer` utility class with tests
+
+### GCGI-950: Attributes
+- Represent attributes as a comma-separated list, instead of individual parameters
+- Add a method to check all attributes are known
+- Define a list of known attributes in `configurable` class; may override in subclasses
+
+### GCGI-951: Dependencies
+- Explicitly represent plugin dependencies with INI parameters
+- Params `depends_configure` and `depends_extract` expect a comma-separated list of component names, which will be checked at runtime
+- Do not define a dependency param at the render step; JSON output from each plugin is expected to be self-contained, so all dependencies should be resolved at the extract step.
+
+### GCGI-955: `specify_params`
+- Each plugin must have a `specify_params` method to define required and optional INI parameters
+- Using an INI parameter not defined in `specify_params` will cause an error
+- Refactor INI and priority handling to enable `specify_params`
+
+### Other
+- Strict substitution for environment variable templates; consistent with HOWTO on wiki
+
 ## v1.0.0-dev0.0.5: 2023-07-04
 
 ### GCGI-946: Versioning for plugins